Throw exceptions in lambda (#73)
Before, exceptions would be logged and then swallowed. This is fine when you are just looking at the logs, but at a certain scale it becomes difficult to find exceptions in cloudwatch. The swallowed exceptions wouldn't appear in the cloudwatch lambda error metrics, so finding exceptions can be difficult at a glance. This PR starts throwing exceptions so that the error metrics get correctly reported, which makes finding exceptions easier.