Add support for HTTP encoding, switch to AWS HTTP API instead of REST (#88)
Add support for HTTP encoding. This is an opt-in feature on HttpServerlessFunctions to decode requests and encode responses. This is useful to avoid the 6MB lambda request / response limit. Currently this only supports gzip encoding.

AWS API Gateway REST APIs were messing up the request decoding - it has a feature where it'll attempt to do the decoding for you. If it succeeds, the request to functions may now exceed the 6MB limit. To bypass this, we need to switch to HTTP APIS. This has a few tradeoffs.
HTTP APIs pros:

3x cheaper for the first tier
More control over requests / responses
Fewer CloudFormation resources required, so less likely to hit the CloudFormation stack resource limit
HTTP API cons:

Unable to enable detailed logging to understand where and why requests are failing
No support for cognito authorizers